Combine CLI source and test projects
