--- Source/Client/CLI.UnitTests/Recipe.json
@@ -1,11 +1,0 @@
-{
-  "name": "Soup.Client.UnitTests",
-  "version": "1.0.0",
-  "type": "executable",
-  "dependencies": [
-    "../Client/"
-  ],
-  "source": [
-    "gen/Main.cpp"
-  ]
-}
