--- Source/Client/CLI.UnitTests/gen/Commands/InstallCommandTests.gen.h
@@ -1,12 +1,0 @@
-#pragma once
-#include "Commands/InstallCommandTests.h"
-
-TestState RunInstallCommandTests() 
- {
-	auto className = "InstallCommandTests";
-	auto testClass = std::make_shared<Soup::Client::UnitTests::InstallCommandTests>();
-	TestState state = { 0, 0 };
-	state += Soup::Test::RunTest(className, "Initialize", [&testClass]() { testClass->Initialize(); });
-
-	return state;
-}
