+++ Source/Client/CLI/Recipe.toml
@@ -16,15 +16,32 @@ RuntimeDependencies = [
 	"LocalUserConfig.json",
 ]
 Source = [
-	"Main.cpp",
+	"Source/Main.cpp",
 ]
 IncludePaths = [
-	"Commands/",
-	"Options/",
+	"Source/Commands/",
+	"Source/Options/",
 ]
 PlatformLibraries = [
 	"OleAut32.lib",
 	"shell32.lib",
 	"User32.lib",
 ]
+DevDependencies = [
+	# ToDO: "../../Extensions/TestBuild/",
+	"Soup.Test.Assert@0.1.7",
+]
+
+	[Tests]
+	Source = [
+		"UnitTests/gen/Main.cpp"
+	]
+	IncludePaths = [
+		"UnitTests/",
+	]
+	PlatformLibraries = [
+		"OleAut32.lib",
+		"shell32.lib",
+		"User32.lib",
+	]
