+++ Source/Client/CLI/UnitTests/gen/Commands/VersionCommandTests.gen.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "Commands/VersionCommandTests.h"
+
+TestState RunVersionCommandTests() 
+ {
+	auto className = "VersionCommandTests";
+	auto testClass = std::make_shared<Soup::Client::UnitTests::VersionCommandTests>();
+	TestState state = { 0, 0 };
+	state += Soup::Test::RunTest(className, "Initialize", [&testClass]() { testClass->Initialize(); });
+
+	return state;
+}
